+++ inject/src/main/scala/com/itiancai/galaxy/inject/Context.scala
@@ -1,10 +1,15 @@
 package com.itiancai.galaxy.inject
 
-import com.itiancai.galaxy.inject.config.ConfigureEnvironment
+import com.itiancai.galaxy.inject.config.{ConfigureEnvironment2, ConfigureEnvironment}
 import org.springframework.beans.factory.BeanFactory
 import org.springframework.beans.factory.config.ConfigurableBeanFactory
 import org.springframework.context.annotation.{Bean, ComponentScan, Configuration, AnnotationConfigApplicationContext}
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer
+import org.springframework.core.env.ConfigurableEnvironment
+import org.springframework.core.io.support.ResourcePatternResolver
+
+import com.twitter.app.{App => TwitterUtilApp}
+
 
 import scala.collection.mutable.ArrayBuffer
 import scala.reflect._
@@ -12,21 +17,16 @@ import scala.reflect._
 
 private[inject] object ContextHolder {
 
-  private[this] val environment = new ConfigureEnvironment()
-
-  private[this] val appContext: AnnotationConfigApplicationContext = {
-    val appContext = new AnnotationConfigApplicationContext()
-    appContext.setEnvironment(environment)
-    appContext
-  }
+  private[this] val appContext: AnnotationConfigApplicationContext = new AnnotationConfigApplicationContext()
 
   private[this] val binder_ = Binder(appContext.getBeanFactory)
   private[this] val injector_ = Injector(appContext.getBeanFactory)
 
   binder_ bind injector_
 
-  private[inject] def injector(annotationClasses: Seq[Class[_]], modules: Seq[Module]): Injector = {
-    appContext.register(annotationClasses: _*)
+  private[inject] def injector(contextConfig: ContextConfig, modules: Seq[Module]): Injector = {
+    appContext.setEnvironment(contextConfig.environment)
+    appContext.register(contextConfig.registerClass : _*)
     modules foreach (module => module.config(binder_))
     appContext.refresh()
     injector_
@@ -71,9 +71,18 @@ private[this] class InjectContextConfig {
 }
 
 
-trait ContextConfig {
+trait ContextConfig extends TwitterUtilApp{
 
   private val registerAnnotationClasses: ArrayBuffer[ClassTag[_]] = ArrayBuffer(classTag[InjectContextConfig])
+
+
+  val DEFAULT_ENV: String = "dev"
+
+  val DEFAULT_ENV_LOCATION: String = ResourcePatternResolver.CLASSPATH_ALL_URL_PREFIX + "/config"
+
+  val rainieEnv = flag("env", DEFAULT_ENV, "rainie current environment, default: dev")
+
+  val configPath = flag("env.config", DEFAULT_ENV_LOCATION, "rainie environment config location path")
 
   def registerClass(): Seq[Class[_]] = {
     registerAnnotationClasses.toSeq map (cl => cl.runtimeClass)
@@ -83,6 +92,13 @@ trait ContextConfig {
     registerAnnotationClasses += classTag[T]
     this
   }
+
+  def environment: ConfigurableEnvironment = {
+    new ConfigureEnvironment(rainieEnv, configPath)
+  }
+
+
+
 
 }