+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [rainie-0.0.4](https://github.com/tftc/rainie/tree/v0.0.3) (2016-06-07)
+
+**New Future**
+
+1. change environment parameter from `-Dgalaxy.env` to twitter Flag, `-env=dev`
+2. can set env config folder path, use parameter `-env.config`, the premix can use 'file:' or 'classpath*:'
+
+## [rainie-0.0.3](https://github.com/tftc/rainie/tree/v0.0.3) (2016-04-27)
+
+**New Future**
+
+1. add Controller for easy to use thrift
+2. change use serverset to use zk
+
+
 ## [rainie-0.0.2](https://github.com/tftc/rainie/tree/v0.0.2) (2016-01-26)
 
 **New Future**